+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Enable internal network names for each service [#638](https://github.com/IN-CORE/pyincore/issues/638)
 - Update production URL references to tools.in-core.org and dev to dev.in-core.org [#634](https://github.com/IN-CORE/pyincore/issues/634)
 
+### Added
+- ML Enabled Galveston CGE [#641](https://github.com/IN-CORE/pyincore/issues/641)
+
 ### Fixed
 - Documentation container tagging error by github action [#631](https://github.com/IN-CORE/pyincore/issues/631)
 - Updated rasterio dependency to the latest to fetch correct GDAL version [#636](https://github.com/IN-CORE/pyincore/issues/636)